two working codes for Euler problems 1 and 2. One initial trial code for Euler problem 2.
The code where I was asked to find the sum of all multiples of 3 or 5 below 1000 was relatively simple. The only mistake I ran into was when using range(). It should be range of 1-1000 inclusive, however range(1,1000)  yields a range of 1-999 inclusive. For the second Euler problem, I ran into a major problem. My initial code was returning a value, but not the correct one. Since it outputted the final new_fib value, or 4000000, I thought there was a problem when I store the new fib number as the original fib number and the new-fib number + original fib number as the new fib number. Before I could confuse myself too much, I decided to combine both expressions into one single expression with commas separating the original fib expression and the new fib expression.