Attempted Euler problem 5 and 7
This work session was almost as bad as the first one. I reattempted Euler problem 5, which I struggled to solve on my very first work session. This time, i had some progress, as I was able to create a helper function that finds the prime factors of all numbers from range(1,n). However, I eventually returned to the same problem I faced on my first attempt, which is extracting the unique prime numbers from the list. After spending a long time thinking, I decided to move on to the next problem instead of wasting time trying meaningless things (like I did in my first work session). However, problem 7 was also a great challenge. In order to solve the problem, I had to find all the prime numbers in a range of numbers from 1 to n. This was almost the exactly same problem I struggled with in problem 5. After spending some time thinking, I thought of utilizing a for loop within a for loop to check if each number was divisible by any number (check if it is not prime). However, for some reason, my code ended up printing all values from 1 - 10001 and repeating that process.