Working on Euler Problem 5, debugging
Today, I made some good progress on Euler Problem 5. I finally thought of a way to sort out the prime factors with the repeated factors in a given list. I created a helper function sort_factors that performs this task. It took me a very long time to write the initial code, as there are multiple loops and if statements involved in the code. However, debugging was even worse. I ran into many problems with an error message telling me how a certain element is not found on the list, or how the list is empty. This was a big problem, since I intended to move on to the next set of factors once one set of factors(list) has no elements within it anymore (For and while loop). I ended up getting rid of the error message. However, the value outputted from my helper function is an empty list, which is not correct.